Move Ethereum bridge methods to their own files
